% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statstable.R
\name{append_default}
\alias{append_default}
\title{Appends default_statistic drop if necessary}
\usage{
append_default(x)
}
\arguments{
\item{x}{statistic list}
}
\value{
statistic list containing a default_statistic
}
\description{
Appends default_statistic drop if necessary
}
\keyword{internal}
