% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion-conc.R
\name{apply_convert_conc}
\alias{apply_convert_conc}
\title{Helper function to reduce a list of conversions parameters}
\usage{
apply_convert_conc(data, args, ...)
}
\arguments{
\item{data}{to convert the units}

\item{args}{the arguments for the conversion. a named list with parameter, to, from as named item}

\item{...}{additional arguments to pass to convert_units}
}
\value{
result of convert_units
}
\description{
Helper function to reduce a list of conversions parameters
}
\keyword{internal}
