% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bind_rows_with_factor_columns}
\alias{bind_rows_with_factor_columns}
\title{bind rows of data frames with factors}
\usage{
bind_rows_with_factor_columns(...)
}
\arguments{
\item{...}{data frames to combine}
}
\value{
combined data frame
}
\description{
\code{\link[dplyr:bind]{dplyr::bind_rows()}} coerces factors to characters. This function use the \code{\link[forcats:fct_c]{forcats::fct_c()}}
to combine the factor columns.
}
\details{
\emph{Important}: It is mandatory that all data frames have the same structure
}
