% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{calculate_CO_h8gl}
\alias{calculate_CO_h8gl}
\title{Calculates CO 8 hour rolling statistics}
\usage{
calculate_CO_h8gl(data, quiet = FALSE)
}
\arguments{
\item{data}{input data in rolf format should contain 10min or 30min data of CO in mg/m3}

\item{quiet}{Instead of stopping if no parameter is found return an empty frame. Default FALSE}
}
\value{
list with h8gl and y1 stats
}
\description{
This function uses openair::rollingMean to calculate the 8 hour rolling Mean. When openair cannot be loaded, an
error is thrown.

Statistics:
\itemize{
\item h8gl left aligned
\item max h8gl > y1
}
}
