% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{calculate_LRV}
\alias{calculate_LRV}
\title{Calculates exceedance of LRV limits}
\usage{
calculate_LRV(input)
}
\arguments{
\item{input}{data in rolf format}
}
\value{
list with m1 and y1 stats
}
\description{
Calculates exceedance of LRV limits
}
