% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{calculate_mass_concentrations}
\alias{calculate_mass_concentrations}
\title{Prepares data for stats calculations}
\usage{
calculate_mass_concentrations(data, keep_ppb = FALSE)
}
\arguments{
\item{data}{input data in rolf format}

\item{keep_ppb}{usually volume concentrations are not used for analyses and are dropped with the exception of NOx.
When FALSE drops parameters with ppb/ppm units except NOx. Default FALSE}
}
\value{
tibble in rolf format containing mass concentrations
}
\description{
Most indicators are calculated from mass concentrations. This functions converts all volume concentrations to mass
concentration with the exception of NOx. The argument keep_ppb allows to keep the volume concentrations
}
