% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{calculate_O3}
\alias{calculate_O3}
\title{Calculates a lot of O3 indicators}
\usage{
calculate_O3(data, quiet = FALSE)
}
\arguments{
\item{data}{input data in rolf format should contain 10min or 30min data of O3 in \u00b5g/m3}

\item{quiet}{Instead of stopping if no parameter is found return an empty frame. Default FALSE}
}
\value{
list with d1, m1, y1 and
}
\description{
calculates the following direct statistics:
h1 > d1: "n>160", "n>180", "n>200", "n>240", "max", "n"
h1 > m1: "n>160", "n>180", "n>200", "n>240", "max", "n"
h1 > y1: "n>160", "n>180", "n>200", "n>240", "max", "n"
d1 > m1: "n>65"
d1 > y1: "n>65"
h1 > m6,m4-m9: 7h mean from 9:00 - 16:00 CET, AOT40 sum ppb value > 40 from 8:00 - 20:00 CET

And following two step statistics (number of days with hour values > xxx):
h1 > m1: O3_nb_d1_h1>160, O3_nb_d1_h1>180, O3_nb_d1_h1>200, O3_nb_d1_h1>240
h1 > y1: O3_nb_d1_h1>160, O3_nb_d1_h1>180, O3_nb_d1_h1>200, O3_nb_d1_h1>240
}
