% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{calculate_O3_summer}
\alias{calculate_O3_summer}
\title{Calculates AOT40 and 7h mean from 9:00 - 16:00 CET from April until September}
\usage{
calculate_O3_summer(data, quiet = FALSE)
}
\arguments{
\item{data}{input data in rolf format should contain 10min or 30min data of O3 in µg/m3}

\item{quiet}{Instead of stopping if no parameter is found return an empty frame. Default FALSE}
}
\value{
tibble in rolf format containing AOT40 and 7h mean
}
\description{
Calculates AOT40 and 7h mean from 9:00 - 16:00 CET from April until September
}
