% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{calculate_stats}
\alias{calculate_stats}
\title{Calculates most common stats}
\usage{
calculate_stats(input)
}
\arguments{
\item{input}{data in rolf format}
}
\value{
list with stats for h1, d1, m1, y1
}
\description{
Calculates the most common used stats from min10 or min30 input values:
\itemize{
\item Averages (wind vector and scalar, RainDur Sum): input > h1, d1, m1, y1
\item min max:  input > d1, m1, y1  &  d1 > m1, y1
\item n: input > y1  &  d1 > y1
\item perc95 (perc02 & perc98 for O3): input > y1
}
}
