% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statstable.R
\name{calculate_statstable}
\alias{calculate_statstable}
\title{Calculates stats described in a table}
\usage{
calculate_statstable(data, statstable, sep = "\\\\s*,\\\\s*",
  keep_input = FALSE, data_thresh = 0.8, max_gap = 10,
  order = c("input", "h1", "h8gl", "d1", "m1", "y1"))
}
\arguments{
\item{data}{input data in rolf format}

\item{statstable}{description of statistics to calculate in table form}

\item{sep}{seperator for combined values in statstable}

\item{keep_input}{should the input data be kept in return list as item input. Default FALSE}

\item{max_gap}{in days. Only used in calculation to y1. Set to NULL to disable usage. Default 10 days}

\item{order}{defines the order of calculation in the from column}

\item{threshold}{minimum data capture threshold 0 - 1.0 to use. Default 0.8}
}
\value{
list with one item per to interval
}
\description{
With various caveats:
\itemize{
\item always the same data threshold for all calculations
\item to h8gl: only mean from h1
\item only to y1 uses max gap
}
}
\section{TODO}{

\itemize{
\item example needed. See Tests for an example usage
\item documentation needed
}
}

