% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{conversion_fun_factory}
\alias{conversion_fun_factory}
\title{Factory to generate conversion functions}
\usage{
conversion_fun_factory(parameter, from, to, ...)
}
\arguments{
\item{parameter}{to convert}

\item{from}{start unit}

\item{to}{target unit}

\item{...}{addional parameters passed to conversion function. see \code{\link[purrr:partial]{purrr::partial()}}, \code{\link[=parts_to_mass]{parts_to_mass()}},
\code{\link[=mass_to_parts]{mass_to_parts()}}}
}
\value{
conversion function with one argument
}
\description{
Factory to generate conversion functions
}
