% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion-conc.R
\name{convert_conc}
\alias{convert_conc}
\alias{convert_conc_multiple}
\title{Convert between volume and mass concentrations}
\usage{
convert_conc(data, parameter, from, to, method = "return", ...)

convert_conc_multiple(data, conversions, method = "return", ...)
}
\arguments{
\item{data}{containing the parameter to convert}

\item{parameter}{to convert}

\item{from}{start unit}

\item{to}{target unit}

\item{method}{one of "return", "append", "replace". "return" only returns the converted data, "append" appends the
converted data at the end, "replace" replaces the original data for the parameter with the converted}

\item{...}{addional parameters passed to conversion function. see \code{\link[purrr:partial]{purrr::partial()}}, \code{\link[=parts_to_mass]{parts_to_mass()}},
\code{\link[=mass_to_parts]{mass_to_parts()}}}

\item{conversions}{tibble with columns paramter, from, to. each row is a conversion}
}
\value{
converted data
}
\description{
This function converts between volume and mass concentration by standard conditions. The constants used are
equivalent to the one used in AIRMO for sites below 1500m a.s.l:\preformatted{R   = 8.314151     [kg * m2 / (s2 * mol * K)]
T   = 20.00        [°C]
p   = 1013.00      [hPa]

mol masses:
CO  = 28.01        [g / mol]
NO  = 30.01        [g / mol]
NO2 = 46.01        [g / mol]
O3  = 48.00        [g / mol]
SO2 = 64.06        [g / mol]
}

It is possible to change this constants via functions during run time for a session. See \code{\link[=convert_set_R]{convert_set_R()}},
\code{\link[=convert_set_mol_masses]{convert_set_mol_masses()}}, \code{\link[=convert_set_conc_lookup]{convert_set_conc_lookup()}}
}
\examples{
min30 <- system.file("extdata", "Zch_Stampfenbachstrasse_min30_2013_Jan.csv",
                     package = "rOstluft.data", mustWork = TRUE)

airmo_min30 <- read_airmo_csv(min30)
convert_conc(airmo_min30, "NO", "µg/m3", "ppb", method = "return")


conversions <- tibble::tribble(
  ~parameter, ~from, ~to,
  "CO", "ppm", "mg/m3",
  "NO", "ppb", "µg/m3",
  "O3", "ppb", "µg/m3",
  "NO2", "ppb", "µg/m3",
  "SO2", "ppb", "µg/m3"
)

min30 <- system.file("extdata", "Zch_Stampfenbachstrasse_min30_2013_Jan.csv",
                     package = "rOstluft.data", mustWork = TRUE)

airmo_min30 <- read_airmo_csv(min30)
airmo_min30_parts <- dplyr::filter(airmo_min30, .data$unit == "ppb" | .data$unit == "ppm")
convert_conc_multiple(airmo_min30_parts, conversions, method = "return")
}
\seealso{
\itemize{
\item \code{\link[=parts_to_mass]{parts_to_mass()}}
\item \code{\link[=mass_to_parts]{mass_to_parts()}}
\item \code{\link[purrr:partial]{purrr::partial()}}
}
}
