% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-seepolizei.R
\name{convert_date_seepolizei}
\alias{convert_date_seepolizei}
\title{Helper function to ensure correct string format}
\usage{
convert_date_seepolizei(x)
}
\arguments{
\item{x}{date as string, POSIXct, POSIXlt or Date Object}
}
\value{
string in format "\%Y-\%m-\%d"
}
\description{
Helper function to ensure correct string format
}
\keyword{internal}
