% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{convert_interval}
\alias{convert_interval}
\title{Converts a AIRMO Zeitfenster to an interval}
\usage{
convert_interval(interval)
}
\arguments{
\item{interval}{string with AIRMO Zeitfenster}
}
\value{
string with valid string for \code{\link[base:seq.Date]{base::seq.Date()}} or \code{\link[lubridate:round_date]{lubridate::floor_date()}}
}
\description{
This function is far from perfect but works with the non offseting, shifting Zeitfenster
}
\keyword{internal}
