% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{convert_unit}
\alias{convert_unit}
\alias{convert_multiple_units}
\title{Convert the unit of a single parameter}
\usage{
convert_unit(data, parameter, from, to, method = "return", ...)

convert_multiple_units(data, conversions, method = "return", ...)
}
\arguments{
\item{data}{containing the parameter to convert}

\item{parameter}{to convert}

\item{from}{start unit}

\item{to}{target unit}

\item{method}{one of "return", "append", "replace". "return" only returns the converted data, "append" appends the
converted data at the end, "replace" replaces the original data for the parameter with the converted}

\item{...}{addional parameters passed to conversion function. see \code{\link[purrr:partial]{purrr::partial()}}, \code{\link[=parts_to_mass]{parts_to_mass()}},
\code{\link[=mass_to_parts]{mass_to_parts()}}}

\item{conversions}{tibble with columns paramter, from, to. each row is a conversion}
}
\value{
converted data
}
\description{
Convert the unit of a single parameter

Converts the units of multiple parameters
}
\examples{
min30 <- system.file("extdata", "Zch_Stampfenbachstrasse_min30_2013_Jan.csv",
                     package = "rOstluft.data", mustWork = TRUE)

airmo_min30 <- read_airmo_csv(min30)
convert_unit(airmo_min30, "NO", "µg/m3", "ppb", method = "return")


conversions <- tibble::tribble(
  ~parameter, ~from, ~to,
  "CO", "ppm", "mg/m3",
  "NO", "ppb", "µg/m3",
  "O3", "ppb", "µg/m3",
  "NO2", "ppb", "µg/m3",
  "SO2", "ppb", "µg/m3"
)

min30 <- system.file("extdata", "Zch_Stampfenbachstrasse_min30_2013_Jan.csv",
                     package = "rOstluft.data", mustWork = TRUE)

airmo_min30 <- read_airmo_csv(min30)
airmo_min30_parts <- dplyr::filter(airmo_min30, .data$unit == "ppb" | .data$unit == "ppm")
convert_multiple_units(airmo_min30_parts, conversions, method = "return")
}
