% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{cut_conversion_data}
\alias{cut_conversion_data}
\title{Cut data for conversion}
\usage{
cut_conversion_data(data, parameter, unit)
}
\arguments{
\item{data}{tibble with input data}

\item{parameter}{parameter for conversion}

\item{unit}{from this unit to another}
}
\value{
named list: $convert parameter for conversion, $others everything else
}
\description{
This function splits off the parameter for conversion
}
\details{
TODO: create an all purpuse function?
}
\keyword{internal}
