% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cut_on_condition}
\alias{cut_on_condition}
\title{Cuts / splits data frame in list based on condition}
\usage{
cut_on_condition(data, condition, mapping)
}
\arguments{
\item{data}{data frame to cut}

\item{condition}{expression used for grouping}

\item{mapping}{names for list}
}
\value{
list
}
\description{
Cuts / splits data frame in list based on condition
}
\examples{
\dontrun{
data <- cut_on_condition(data, end_interval < endtime, c("TRUE" = "overlaps", "FALSE" = "complete"))
}
}
\keyword{internal}
