% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{detect_interval}
\alias{detect_interval}
\title{Detects the interval of the data data}
\usage{
detect_interval(data)
}
\arguments{
\item{data}{input data in rolf format}
}
\value{
interval as character string
}
\description{
Throws error if multiple Intervals are detected
}
\keyword{internal}
