% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver_rds_local.R
\docType{class}
\name{driver_rds_local}
\alias{driver_rds_local}
\title{Local rds store}
\usage{
driver_rds_local(name, path = NULL, read.only = TRUE)
}
\arguments{
\item{name}{name of the store}

\item{path}{optional path to create the store under.
Defaults to \link[rappdirs:user_data_dir]{rappdirs::user_data_dir(appname = name, appauthor = "rOstluft")}}

\item{read.only}{read only store. disable put, if false and the store doesn't exist, the store will be initiated}
}
\value{
R6 driver_rds_local class Object
}
\description{
Local rds store
}
\section{Fields}{

\describe{
\item{\code{name}}{name of the store}

\item{\code{path}}{root of the store}

\item{\code{rds_path}}{root of all chunks}

\item{\code{rds_content}}{path to the rds file containing statistics of store content}

\item{\code{read.only}}{flag for read.only usage of store}
}}

\section{Methods}{

\code{$type()} returns type of store as string

\code{$get(station, intervall, year, parameters = NULL)} get data from the store. see examples for more information

\code{$put(data)} puts the data into the store. Stops if store is read only

\code{$get_content()} get n data points grouped by intervall, station, parameter, year in the store

\code{$get_list_of_chunks()} get list of all chunks

\code{$get_chunk_path(intervall, station, year)} get the path of chunk from grouping variables

\code{$decode_chunk_filename(chunk_path)} decode base64 filename of chunk

\code{$destroy(confirmation)} removes all files under path from the file system if "DELETE" is supplied as
confirmation
}

\examples{
## init store, creates directory if necessary
store <- rOstluft::driver_rds_local("example_rOstluft", read.only = FALSE)

## read data from airmo export und put into the store
fn <- system.file("extdata", "Zch_Stampfenbachstrasse_2010-2014.csv",
                  package = "rOstluft.data", mustWork = TRUE)
df <- rOstluft::read_airmo_csv(fn)
store$put(df)

## get all data min30 for 2011 and 2012
store$get("Zch_Stampfenbachstrasse", "min30", 2011:2012)

## get only data for O3
store$get("Zch_Stampfenbachstrasse", "min30", 2011:2012, "O3")

## get NOx data from multiple stations
store$get(c("Zch_Stampfenbachstrasse", "Zch_Rosengartenstrasse"), "min30",
          2014, c("NOx", "NO", "NO2"))

## get n data points grouped by intervall, station, parameter, year in the store
store$get_content()

## get list of all chunks
store$get_list_of_chunks()

## get the path of chunk from grouping variables
store$get_chunk_path("min30", "Zch_Stampfenbachstrasse", 2012)

## decode base64 filename of chunk
store$decode_chunk_filename("bWluMzAtWmNoX1N0YW1wZmVuYmFjaHN0cmFzc2UtMjAxMg.rds")

## destroy store (careful removes all files on the disk)
store$destroy("DELETE")

}
