% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statstable.R
\name{expand_inputs}
\alias{expand_inputs}
\title{Adds statistics for parameters in inputs}
\usage{
expand_inputs(x, inputs)
}
\arguments{
\item{x}{statistic list}

\item{inputs}{parameters in input}
}
\value{
expanded statistic list
}
\description{
Adds statistics for parameters in inputs
}
\keyword{internal}
