% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statstable.R
\name{expand_statstable}
\alias{expand_statstable}
\title{Expands compact statstable to a table with one stats per row}
\usage{
expand_statstable(statstable, sep = "\\\\s*,\\\\s*")
}
\arguments{
\item{statstable}{compact statstable}

\item{sep}{seperator of combined values}
}
\value{
expanded statstable
}
\description{
Expands compact statstable to a table with one stats per row
}
