% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmean-full-intervall.R
\name{fill_wmean}
\alias{fill_wmean}
\title{Fills (upsample) a data frame for wmean}
\usage{
fill_wmean(data, ..., interval)
}
\arguments{
\item{data}{input data frame}

\item{...}{grouping columns}

\item{interval}{for upsampling}
}
\value{
upsampled data frame
}
\description{
Fills (upsample) a data frame for wmean
}
\keyword{internal}
