% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_keep_list}
\alias{filter_keep_list}
\title{Keep matching rows}
\usage{
filter_keep_list(data, filter_list)
}
\arguments{
\item{data}{tibble to filter}

\item{filter_list}{named list, name = column to match}
}
\value{
filtered tibble
}
\description{
keep all rows with matching all name (=column)/value combination from filter_list.
}
\keyword{internal}
