% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-hysplit.R
\docType{class}
\name{format_hysplit}
\alias{format_hysplit}
\title{Format defintion for hysplit back trajectories}
\usage{
format_hysplit(tz = "Etc/GMT")
}
\arguments{
\item{tz}{Unused at the moment}
}
\value{
R6 class object of format_hysplit
}
\description{
This formats describes the form of back trajectories calculated by the NOAA HYSPLIT model.
}
\section{TODO}{

Write the Documentation ...
}

