% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic.R
\name{gap_wrapper_function}
\alias{gap_wrapper_function}
\title{wrapper function for maximum number of consecutive NA values}
\usage{
gap_wrapper_function(statistic_fun, max_gap)
}
\arguments{
\item{statistic_fun}{statistic function to apply if max gap criterium is meet}

\item{max_gap}{maxium Number of consecutive NA values}
}
\value{
wrapped function
}
\description{
wrapper function for maximum number of consecutive NA values
}
