% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statstable.R
\name{get_list}
\alias{get_list}
\title{convert table to resample compatible list}
\usage{
get_list(table)
}
\arguments{
\item{table}{statstable}
}
\value{
list
}
\description{
convert table to resample compatible list
}
\keyword{internal}
