% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statstable.R
\name{get_statistics_from_table}
\alias{get_statistics_from_table}
\title{Create resample compatible statistic lists}
\usage{
get_statistics_from_table(table, from, inputs = NULL)
}
\arguments{
\item{table}{statstable}

\item{from}{interval to generate statistic list}

\item{inputs}{parameters in input}
}
\value{
list with statistic list for every to
}
\description{
Could be exported for user to debug his table
}
\keyword{internal}
