% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getElement2}
\alias{getElement2}
\title{Get element or default value from Object}
\usage{
getElement2(object, element_name, default_value = NULL)
}
\arguments{
\item{object}{object from which to get element}

\item{element_name}{a string matched to the names of the object}

\item{default_value}{return value if element not found in object}
}
\value{
element or default value
}
\description{
Get element or default value from Object
}
\keyword{internal}
