% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-import.R
\name{import_directory}
\alias{import_directory}
\title{Imports directory into store}
\usage{
import_directory(store, path, read_function, glob = NULL,
  regexp = NULL, recursive = TRUE, ...)
}
\arguments{
\item{store}{instance of the store to import data into}

\item{path}{to directory containing the data files}

\item{read_function}{function with argument file, ... is passed to this function to supply additionals arguments}

\item{glob}{see \code{\link[fs:dir_ls]{fs::dir_ls()}}}

\item{regexp}{see \code{\link[fs:dir_ls]{fs::dir_ls()}}}

\item{recursive}{see \code{\link[fs:dir_ls]{fs::dir_ls()}}}

\item{...}{passed to the read_function}
}
\description{
Imports directory into store
}
\examples{
format <- rOstluft::format_rolf()
store <- rOstluft::storage_local_rds("example_rOstluft", format, read.only = FALSE)
path <- system.file("extdata", package = "rOstluft.data")

rOstluft::import_directory(store, path, rOstluft::read_airmo_csv)

store$get_content()

store$destroy("DELETE")

}
