% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{mass_to_parts}
\alias{mass_to_parts}
\title{Converts mass to volume concentration}
\usage{
mass_to_parts(values, mol_mass, temperature = 20, pressure = 1013.25,
  ...)
}
\arguments{
\item{values}{mass concentration}

\item{mol_mass}{in kg/mol}

\item{temperature}{in °C. Default 20 °C}

\item{pressure}{in hPa. Default 1013.25 hPa}

\item{...}{catch additional arguments to keep R happy}
}
\value{
volume concentration
}
\description{
Converts mass to volume concentration
}
