% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_content}
\alias{merge_content}
\title{Merge two data frame with store content statistic into one}
\usage{
merge_content(x, y)
}
\arguments{
\item{x}{data frame}

\item{y}{data frame}
}
\value{
combined data frame
}
\description{
Combines the two data frames. Defaulting to data frame \code{y}.
}
\keyword{internal}
