% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_rOstluft_longformat}
\alias{merge_rOstluft_longformat}
\title{Merge two data frames in rOstluft long format into one}
\usage{
merge_rOstluft_longformat(x, y)
}
\arguments{
\item{x}{data.frame in rOstluft long format}

\item{y}{data.frame in rOstluft long format}
}
\value{
merged data frame
}
\description{
Combines the two data frames. Defaulting to data frame \code{y}.
}
