% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{pad}
\alias{pad}
\alias{pad_serie}
\title{Pad data}
\usage{
pad(data, start.date = NULL, end.date = NULL, drop.last = FALSE)

pad_serie(serie, start.date = NULL, end.date = NULL,
  drop.last = FALSE)
}
\arguments{
\item{data}{tibble in rOstluft long format}

\item{start.date}{optional start date for padding. Default min date in series}

\item{end.date}{optional end date for padding. Default max date in series}

\item{drop.last}{optional drop the last added time point by padding. Useful when
resampling and end.date is the first time point of the new interval}

\item{serie}{a tibble in rOstluft long format containing exactly one serie}
}
\value{
tibble with padded data
}
\description{
Function to pad out missing time points
}
