% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{pad_input}
\alias{pad_input}
\title{Pads data data to complete years}
\usage{
pad_input(data)
}
\arguments{
\item{data}{input data in rolf format}
}
\value{
padded data in rolf format
}
\description{
TODO: Export? could be useful for an user
}
\keyword{internal}
