% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad.R
\name{pad_year}
\alias{pad_year}
\title{Pads data data to complete years}
\usage{
pad_year(data)
}
\arguments{
\item{data}{input data in rolf format}
}
\value{
padded data in rolf format
}
\description{
Pads data data to complete years
}
