% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{parts_to_mass}
\alias{parts_to_mass}
\title{Converts mass to volume concentration}
\usage{
parts_to_mass(values, mol_mass, temperature = 20, pressure = 1013, ...)
}
\arguments{
\item{values}{volume concentration}

\item{mol_mass}{in g/mol}

\item{temperature}{in °C. Default 20 °C}

\item{pressure}{in hPa. Default 1013.25 hPa}

\item{...}{catch additional arguments to keep R happy}
}
\value{
mass concentration
}
\description{
Converts mass to volume concentration
}
