% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-airmo.R
\name{read_airmo_webexport}
\alias{read_airmo_webexport}
\title{Reads an AIRMO Export in CSV Webexport Format}
\usage{
read_airmo_webexport(
  fn,
  encoding = "latin1",
  tz = "Etc/GMT-1",
  time_shift = NULL,
  na.rm = TRUE
)
}
\arguments{
\item{fn}{CSV file name, can be relative to current directory or absolut}

\item{encoding}{CSV file encoding. Default "latin1"}

\item{tz}{time zone of date field. Be carefull Etc/GMT + == -. Default "Etc/GMT-1"}

\item{time_shift}{a lubridate period to add to the time. Default NULL}

\item{na.rm}{remove na (empty) values. Default TRUE}
}
\value{
data frame in rOstluft long format
}
\description{
Reads the data from a AIRMO Webexport CSV file and returns it as a data frame in rOstluft long format
}
\examples{
input <- rOstluft.data::f("ol_nox_covid19_2020.csv")
read_airmo_webexport(input)
}
\seealso{
\code{\link[lubridate:period]{lubridate::period()}} - Create or parse period objects

\link[base:timezones]{base::timezones} - Information about time zones in R
}
