% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-ethz-iac.R
\name{read_ethz_iac}
\alias{read_ethz_iac}
\title{Read an IAC/ETHZ meteo station file}
\usage{
read_ethz_iac(x, tz = "Etc/GMT-1", site = NULL, na.rm = TRUE,
  encoding = "UTF-8")
}
\arguments{
\item{x}{Either a path to a file, a connection, or literal data. see \code{\link[readr:read_delim]{readr::read_delim()}}}

\item{tz}{of the output data. Default "Etc/GMT-1"}

\item{site}{Character string specifying the site of the meteo station. Usally the site will be autodetected from the
header. But if autodection fails or an overwrite is needed set site. Default NULL}

\item{na.rm}{remove na values. Default TRUE}

\item{encoding}{encoding of the data file. Default = "UTF-8"}
}
\value{
tibble in rOstluft long format structure
}
\description{
Reads files from meteo station measurements (10 minute time resolution) by the institute for atmosphere and climety
at ETH Zürich. Currently, two stations in Zürich are under long-term operation: CHN-Gebäude & Hönggerberg (see
\link{http://www.iac.ethz.ch/the-institute/weather-stations.html}(Weather Stations)).
}
\details{
The files are daily files, encoded in utf-8. Timezone is GMT.
}
