% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-smn.R
\name{read_meteoschweiz_smn}
\alias{read_meteoschweiz_smn}
\title{Read a MeteoSchweiz SwissMetNet export file}
\usage{
read_meteoschweiz_smn(x, timezone = "Etc/GMT", encoding = "UTF-8",
  time_format = "\%Y\%m\%d\%H\%M")
}
\arguments{
\item{x}{Either a path to a file, a connection, or literal data. see \code{\link[readr:read_delim]{readr::read_delim()}}}

\item{timezone}{time zone of date fields. Be carefull Etc/GMT '+' actually signifies '-' and vice versa, e.g.
'Etc/GMT-1' = UTC + 1. Default: "Etc/GMT"}

\item{encoding}{file encoding. Default: "UTF-8"}

\item{time_format}{of data in file, different for 10min and 1H}
}
\value{
tibble in rOstluft long format structure
}
\description{
Read a MeteoSchweiz SwissMetNet export text file (such as '') and restructure the data into long format
}
\seealso{
\link[base:timezones]{base::timezones} - Information about time zones in R

\link[base:strptime]{base::strptime} - Information about time format strings in R
}
