% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-nabel.R
\name{read_nabel_txt}
\alias{read_nabel_txt}
\title{Read a file from Nabel}
\usage{
read_nabel_txt(fn, encoding = "latin1", tz = "Etc/GMT-1",
  interval = NULL, time_shift = NULL, na.rm = TRUE)
}
\arguments{
\item{fn}{path to input file}

\item{encoding}{encoding of the data file. Default = "latin1"}

\item{tz}{of the output data. Default "Etc/GMT-1"}

\item{interval}{a lubridate period to add to the time. Default NULL}

\item{time_shift}{optional interval of the data. Use if auto detect fails. Default NULL. If used it is necessary to
define time_shift manuelly. lubridate::period(0) can be used for no shifting}

\item{na.rm}{remove na values. Default TRUE}
}
\value{
tibble in rolf format
}
\description{
Read a file from Nabel
}
