% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-seepolizei.R
\name{read_seepolizei}
\alias{read_seepolizei}
\title{Reads historic data files  of the Tecson meteo station placed at Tiefenbrunnen and Mythenquai water police station.}
\usage{
read_seepolizei(
  fn,
  site = c("ZH_Mythenquai", "ZH_Tiefenbrunnen"),
  tz = "Etc/GMT-1",
  time_shift = NULL,
  na.rm = TRUE
)
}
\arguments{
\item{fn}{file name containing historic data}

\item{site}{measurement site of data. Should be "ZH_Mythenquai" or "ZH_Tiefenbrunnen"}

\item{tz}{Output time zone. The file content is in CET. Default "Etc/GMT-1"}

\item{time_shift}{a lubridate period to add to the time. Default NULL}

\item{na.rm}{remove na (empty) values. Default TRUE}
}
\value{
data frame in rOstluft long format
}
\description{
Reads historic data files  of the Tecson meteo station placed at Tiefenbrunnen and Mythenquai water police station.
}
\examples{
input <- rOstluft.data::f("messwerte_tiefenbrunnen_2007-2019.csv")
data <- read_seepolizei(input, "ZH_Tiefenbrunnen")

# normalize data with meta informations
meta_data <- readRDS(rOstluft.data::f("meta_seepolizei.rds"))

# add units
data <- meta_apply(
  data = data,
  meta = meta_data,
  data_src = "parameter",
  data_dest = "unit",
  meta_key = "parameter_original",
  meta_val = "unit",
  mode = "strict"
)

# rename parameters to Ostluft Conventions
data <- meta_apply(
  data = data,
  meta = meta_data,
  data_src = "parameter",
  data_dest = "parameter",
  meta_key = "parameter_original",
  meta_val = "parameter",
  mode = "strict"
)

# show normalized data
data
}
