% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-smn.R
\name{read_smn_multiple}
\alias{read_smn_multiple}
\title{Reads a file from the SwissMetNet of MeteoSwiss containing multiple Exports}
\usage{
read_smn_multiple(fn, as_list = FALSE, encoding = "UTF-8", ...)
}
\arguments{
\item{fn}{path to input file}

\item{as_list}{return a list with each part as element rather than one concatenated tibble}

\item{encoding}{encoding of the data file. Default = "UTF-8"}

\item{...}{parameters passed to \code{\link[=read_smn]{read_smn()}}}
}
\value{
tibble or list of tibbles in rOstluft long format structure
}
\description{
One peculiarity of MeteoSwiss is to provide multiple exports concatenated in one file, starting with a line
containing a single space and seperated by two newlines. This function reads the whole file into the memory, splits
the part and applies the read function to each part. Is the file too big for the available memory, the function
\code{\link[=split_smn]{split_smn()}} can split the file into multiple files each containing one part. On Linux the commandline tool csplit
is another probably faster tool.
}
\examples{
fn <- system.file("extdata", "smn_multi.txt", package = "rOstluft.data", mustWork = TRUE)
read_smn_multiple(fn, as_list = TRUE)

}
\seealso{
\itemize{
\item \code{\link[=read_smn]{read_smn()}}
\item \code{\link[=split_smn]{split_smn()}}
\item \href{http://man7.org/linux/man-pages/man1/csplit.1.html}{csplit} - split a file into sections determined by context
lines
}
}
