% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample}
\alias{resample}
\alias{resample_series}
\alias{resample_wind}
\title{Resampling data}
\usage{
resample(data, statistic = "mean", new_interval, data.thresh = NULL,
  start.date = NULL, end.date = NULL, drop.last = FALSE,
  rename.parameter = FALSE, percentile = 95)

resample_series(serie, statistic = "mean", new_interval = "d1",
  at = "value", data.thresh = NULL, start.date = NULL,
  end.date = NULL, drop.last = FALSE, rename.parameter = FALSE,
  percentile = 95)

resample_wind(data, statistic, new_interval = "d1", data.thresh = NULL,
  start.date = NULL, end.date = NULL, drop.last = FALSE)
}
\arguments{
\item{data}{A tibble in rOstluft long format}

\item{statistic}{Statistical method to apply when aggregating the data; default is the mean.
Can be one of “mean”, “max”, “min”, “median”, “frequency”, “sd”, “percentile”. Note that “sd”
is the standard deviation, “frequency” is the number (frequency) of valid records in the period
and “data.cap” is the percentage data capture. “percentile” is the percentile level (\%) between
0-100, which can be set using the “percentile” argument.
Or a function with one argument expecting a vector.
Or a list with parameter as name and the statistical method as value (function or name of method)}

\item{new_interval}{New interval. Must be longer than actual interval (not checked)}

\item{data.thresh}{minimum data capture threshold in \% to use}

\item{start.date}{optional start date for padding. Default min date in series floored to the new interval}

\item{end.date}{optional end date for padding. Default max date in series ceiled to the new interval}

\item{drop.last}{optional drop the last added time point by padding. Default False, true if no end.date
provided and max date != ceiled max date.}

\item{percentile}{The percentile level in \% used when statistic = "percentile". The default is 95.}

\item{serie}{a tibble in rOstluft long format containing exactly one serie}
}
\value{
tibble with resampled data

tibble with resampled series

tibble with resampled wind data
}
\description{
Aggregate data by different time periods. Following this simple steps:
\itemize{
\item split data in series
\itemize{
\item pad data serie (needed for calculation of capture threshold)
\item group serie by new interval with \code{\link[lubridate:floor_date]{lubridate::floor_date()}}
\item apply statistical method or user provides function (user can provide list per parameter)
}
\item combine resampled series
}
}
\section{TODO Wind}{

Code optimieren ?
sicherstellen, dass alle Voraussetzungen erfüllt sind (stopifnot() etc) und korrektes bind_rows mit Faktoren
(Deine bind_rows_with... spuckt die Zahl aus, nicht einen Faktor, in meinem Fall)
}

\keyword{internal}
