% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample_series}
\alias{resample_series}
\title{resampling a serie}
\usage{
resample_series(serie, statistic = "mean", new_interval = "d1",
  data.thresh = NULL, start.date = NULL, end.date = NULL,
  drop.last = FALSE, rename.parameter = FALSE, percentile = 0.95,
  max_gap = NULL)
}
\arguments{
\item{serie}{a tibble in rOstluft long format containing exactly one serie}
}
\value{
tibble with resampled series
}
\description{
resampling a serie
}
\keyword{internal}
