% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{rolf_apply_meta}
\alias{rolf_apply_meta}
\title{apply meta data to rolf data}
\usage{
rolf_apply_meta(data, meta, parameter_orig, parameter = NULL,
  unit = NULL, mode = "strict", replacement_parameter = NULL,
  replacement_unit = NULL)
}
\arguments{
\item{data}{input data}

\item{meta}{meta data}

\item{parameter_orig}{column in meta for parameter and unit remapping}

\item{parameter}{source column in meta for replacement of parameter_orig values in data for parameter remapping}

\item{unit}{source column in meta for replacement of parameter_orig values in data for unit remapping}

\item{mode}{one of strict, drop, keep, replace}

\item{replacement_parameter}{named list of replacement values for missing parameter_orig values in meta for
parameter remapping}

\item{replacement_unit}{named list of replacement values for missing parameter_orig values in meta for
unit remapping}
}
\value{
in rOstluft long format structure
}
\description{
See Tests for examples
}
\section{TODO}{

write one universel function: shouldn't be too hard with nse:
\itemize{
\item source and target columns in data (parameter > parameter,  parameter > unit)
\item mapping columns in meta for source ( parameter_original > parameter, parameter_original > unit)
\item the worst will be the user messages
}

just started the wrong way. Wanted to do both remappings in one go ...
}

