% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openair.R
\name{rolf_to_openair}
\alias{rolf_to_openair}
\title{Converts data from rolf format to an openair compatible Format}
\usage{
rolf_to_openair(data, as_list = FALSE, interval = NULL,
  keep_ppb = FALSE, keep_interval = FALSE, ws = "WVv", wd = "WD")
}
\arguments{
\item{data}{in rolf format}

\item{as_list}{optional, if TRUE returns a list a tibble for each parameter and keeps the unit information. Handy
for applying functions with purrr.}

\item{interval}{optional filter for interval}

\item{keep_ppb}{usually volume concentrations are not used for analyses and are dropped. When TRUE keep parameters
with ppb/ppm units. Default FALSE}

\item{keep_interval}{keep the "interval" column. Default FALSE}

\item{ws}{renames this parameter to "ws". Default "WVv". Set ws = NULL to disable renaming}

\item{wd}{renames this parameter to "wd". Default "WD". Set wd = NULL to disable renaming}
}
\value{
tibble in openair compatible format
}
\description{
\link[openair:openair]{openair::openair} provides Tools for the analysis of air pollution data. rOstluft focus lies in providing data
from different sources to the user. The openair data format is described in \link[openair:mydata]{openair::mydata}. Basically it is a
wide format with a field "date" as POSIXct or Date and the parameters as columns. Another convention are the fields
"ws" for wind speed and "wd" for wind direction. Although usually the functions allows passing of the names from the
wind fields with the arguments wd and ws. One disadvantage of the wide format is the loss of unit information.
Normally all parameters are in expressed in mass terms. See \code{\link[openair:importKCL]{openair::importKCL()}} and \code{\link[openair:importAURN]{openair::importAURN()}}.
This functions saves the parameter <> unit information in the attribute "units". unfortunately attributes tend to be
lost by data wrangling functions. But an avid user might save the information.
}
\details{
Some functions, for example \link[openair:timeAverage]{openair::timeAverage}, have the argument "type". This argument allows to specifiy
columns as grouping columns. However, if type is not supplied grouping variables (character or factor) will be
dropped. Most other functions need a mandatory assignment of the argument pollutant.
}
\examples{
h1 <- system.file("extdata", "Zch_Stampfenbachstrasse_h1_2013_Jan.csv",
                   package = "rOstluft.data", mustWork = TRUE)

airmo_h1 <- read_airmo_csv(h1)
rolf_to_openair(airmo_h1)

}
