% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openair.R
\name{rolf_to_openair_single}
\alias{rolf_to_openair_single}
\title{Converts one parameter in a rolf data frame to an openair compatible format}
\usage{
rolf_to_openair_single(
  data,
  parameter = NULL,
  unit = NULL,
  interval = NULL,
  keep_interval = FALSE
)
}
\arguments{
\item{data}{containing the parameter in rolf format}

\item{parameter}{to pluck}

\item{unit}{Optional: unit of the parameter to pluck. Default NULL}

\item{interval}{Optional: interval of the parameter to pluck. Default NULL}

\item{keep_interval}{Optional: keep interval column. Default FALSE}
}
\value{
tibble with one parameter in openair compatible format including unit, site and optional interval column
}
\description{
This function plucks the data for one parameter. If multiple combinations of parameter, unit and interval exists it
is necessary to specify which unit and/or interval to pluck.
}
\examples{
h1 <- system.file("extdata", "Zch_Stampfenbachstrasse_h1_2013_Jan.csv",
                   package = "rOstluft.data", mustWork = TRUE)

airmo_h1 <- read_airmo_csv(h1)
rolf_to_openair_single(airmo_h1, "NO2", unit = "µg/m3", keep_interval = TRUE)

}
