% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rOstluft.R
\docType{package}
\name{rOstluft}
\alias{rOstluft}
\title{rOstluft: data store and utility functions for Ostluft}
\description{
This package provides function to quickly load data and merge data from mulitple sources.
}
\keyword{internal}
