% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-smn.R
\name{split_smn}
\alias{split_smn}
\title{Splits a file from the SwissMetNet of MeteoSwiss containing multiple Exports}
\usage{
split_smn(fn, out_dir = NULL, suffix = "\%03d.part", encoding = "UTF-8")
}
\arguments{
\item{fn}{path to input file}

\item{out_dir}{Output directory for parts}

\item{suffix}{Added suffix to file name as sprintf format string. Gets file counter as argument.}

\item{encoding}{Encoding of the input file}
}
\description{
Splits a file from the SwissMetNet of MeteoSwiss containing multiple Exports
}
