% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic.R
\name{statistic_mean}
\alias{statistic_mean}
\alias{statistic_median}
\alias{statistic_sd}
\alias{statistic_frequency}
\alias{statistic_sum}
\alias{statistic_max}
\alias{statistic_min}
\alias{statistic_data.cap}
\alias{get_statistic_percentile}
\title{statistic functions}
\usage{
statistic_mean(x)

statistic_median(x)

statistic_sd(x)

statistic_frequency(x)

statistic_sum(x)

statistic_max(x)

statistic_min(x)

statistic_data.cap(x)

get_statistic_percentile(percentile)
}
\description{
statistic functions
}
\keyword{internal}
