% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{statstable_default}
\alias{statstable_default}
\title{Default Statstable}
\usage{
statstable_default(expand = FALSE)
}
\arguments{
\item{expand}{if TRUE statstable is expanded}
}
\value{
statstable
}
\description{
Default Statstable
}
