% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statstable.R
\name{statstable_expand}
\alias{statstable_expand}
\title{Expands compact statstable to a table with one stats per row}
\usage{
statstable_expand(statstable, sep = "\\\\s*,\\\\s*")
}
\arguments{
\item{statstable}{compact statstable}

\item{sep}{seperator of combined values}
}
\value{
expanded statstable
}
\description{
For each comma seperated value in a cell a new row is added.
}
\examples{
statstable <- tibble::tribble(
  ~parameter, ~statistic, ~from, ~to,
  "O3, NO2", "min, max", "h1, d1", "m1, y1"
)
statstable_expand(statstable)

}
\seealso{
\itemize{
\item \code{tidyr::seperate_rows()} is used to expand the table. One call per column.
}
}
