% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{storage_chunk_grouping}
\alias{storage_chunk_grouping}
\title{Chunks and store data}
\usage{
storage_chunk_grouping(groups, storage)
}
\arguments{
\item{storage}{to store data into}

\item{data}{for chunking}
}
\value{
tibble with n = count of data points per chunk per series
}
\description{
Chunks and store data
}
\keyword{internal}
