% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{storage_chunk_nest}
\alias{storage_chunk_nest}
\title{Chunks and store data with additionals calculated}
\usage{
storage_chunk_nest(data, storage)
}
\arguments{
\item{data}{for chunking}

\item{storage}{to store data into}
}
\value{
tibble with n = count of data points per chunk per series
}
\description{
Chunks and store data with additionals calculated
}
\keyword{internal}
