% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_local_rds.R
\docType{class}
\name{storage_local_rds}
\alias{storage_local_rds}
\title{Local rds storage}
\usage{
storage_local_rds(name, format, path = NULL, read.only = TRUE)
}
\arguments{
\item{name}{name of the store}

\item{format}{data format of the store}

\item{path}{optional path to create the store under.
Defaults to \link[rappdirs:user_data_dir]{rappdirs::user_data_dir(appname = name, appauthor = "rOstluft")}}

\item{read.only}{read only store. disable put, if false and the store doesn't exist, the store will be initiated}
}
\value{
R6 class object of r6_storage_local_rds
}
\description{
A local storage with the file format rds. The format defines the data chunks per file.
}
\section{Fields}{

\describe{
\item{\code{name}}{name of the store}

\item{\code{format}}{data format of the store}

\item{\code{path}}{root of the store}

\item{\code{rds_path}}{root of all chunks}

\item{\code{rds_content}}{path to the rds file containing statistics of store content}

\item{\code{read.only}}{flag for read.only usage of store}
}}

\section{Methods}{


\code{$get(filter=NULL, ...)} get data from the store. The name of the arguments depend on the format. The filter
argument is applied to each chunk.

\code{$put(data)} puts the data into the store. Stops if store is read only

\code{$get_content()} returns a tibble with the amount of data points per chunk per series

\code{$list_chunks()} get list of all chunks

\code{$destroy(confirmation)} removes all files under path from the file system if "DELETE" is supplied as
confirmation
}

\examples{
## init store, creates directory if necessary
rolf <- rOstluft::format_rolf()
store <- rOstluft::storage_local_rds("example_rOstluft", rolf, read.only = FALSE)

## read data from airmo export und put into the store
fn <- system.file("extdata", "Zch_Stampfenbachstrasse_2010-2014.csv",
                  package = "rOstluft.data", mustWork = TRUE)
df <- rOstluft::read_airmo_csv(fn)
store$put(df)

fn <- system.file("extdata", "Zch_Rosengartenstrasse_2010-2014.csv",
                  package = "rOstluft.data", mustWork = TRUE)
df <- rOstluft::read_airmo_csv(fn)
store$put(df)

## get all data min30 for 2011 and 2012
store$get(site="Zch_Stampfenbachstrasse", interval="min30", year=2011:2012)

## get only data for O3
store$get(year=2011:2012, site="Zch_Stampfenbachstrasse", interval="min30", filter = parameter == "O3")

## get NOx data from multiple stations
store$get(site = c("Zch_Stampfenbachstrasse", "Zch_Rosengartenstrasse"), interval = "min30",
          year = 2014, filter = parameter \%in\% c("NOx", "NO", "NO2"))

## get n data points grouped by intervall, station, parameter, year in the store
store$get_content()

## get list of all chunks
store$list_chunks()

## destroy store (careful removes all files on the disk)
store$destroy("DELETE")

}
