% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stores.R
\name{store_aqmet}
\alias{store_aqmet}
\alias{store_hysplit}
\alias{store_aqmet_public}
\title{Predefined stores}
\usage{
store_aqmet(name = "aqmet")

store_hysplit(name = "hysplit")

store_aqmet_public(name = "aqmet_public")
}
\arguments{
\item{name}{of the store (to influence caching location)}
}
\value{
store
}
\description{
\itemize{
\item \code{store_aqmet()} - Ostluft S3 storage containing airquality and meteo data
\item \code{store_aqmet_public()} - Public Ostluft S3 storage containing airquality and meteo data.
\item \code{store_hysplit()} - Ostluft S3 storage containing NOAA Hysplit trajectories
}
}
\section{store_aqmet_public}{

This store contains data from following sources:
\subsection{\href{https://www.ostluft.ch/}{Ostluft}}{

Ostluft is a inter cantonal measurement network of eastern switzerland and the Principality of Liechtenstein.
}

\subsection{\href{http://www.iac.ethz.ch/the-institute/weather-stations.html}{ETHZ IAC}}{

Institute for Atmosheric and Climate Science (IAC), ETH Zurich

This data can be used freely for scientific analysis.
The user agrees to cite IAC ETH Zurich as source of the data and to consult IAC ETH Zurich before data is used for publication.

Please note that we do not take any guarantee for the correctness and completeness of the available data.
}

\subsection{\href{https://www.bafu.admin.ch/bafu/en/home/topics/air/state/data/national-air-pollution-monitoring-network--nabel-.html}{National Air Pollution Monitoring Network (NABEL)}}{

The National Air Pollution Monitoring Network (NABEL) measures air pollution at 16 locations in Switzerland.
The stations are distributed throughout the country and monitor pollution at typical locations (e.g.
city-centre streets, residential areas, rural stations). The monitoring network has commenced operations in stages
since 1979 and is operated by the Federal office for the environment and Empa.
}

\subsection{\href{https://data.stadt-zuerich.ch/dataset/sid_wapo_wetterstationen}{Water police Zuerich}}{

The water police Zuerich operates two meteo station. The locations are Tiefenbrunnen and Mythenquai.
}
}

