% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-projection.R
\name{transform}
\alias{transform}
\alias{transform_LV95_to_WSG84}
\alias{transform_WSG84_to_LV95}
\alias{transform_LV03_to_WSG84}
\alias{transform_WSG84_to_LV03}
\title{Transform between Coordinate Reference Systems}
\usage{
transform(data, coord, in_crs, out_crs, append = TRUE)

transform_LV95_to_WSG84(data, coord = c(lon = "x", lat = "y"),
  append = TRUE)

transform_WSG84_to_LV95(data, coord = c(x = "lon", y = "lat"),
  append = TRUE)

transform_LV03_to_WSG84(data, coord = c(lon = "x", lat = "y"),
  append = TRUE)

transform_WSG84_to_LV03(data, coord = c(x = "lon", y = "lat"),
  append = TRUE)
}
\arguments{
\item{data}{input data}

\item{coord}{mappping of columns for Conversion}

\item{in_crs}{source coordinate reference system as \link[sp:CRS]{sp::CRS}}

\item{out_crs}{target coordinate reference system as \link[sp:CRS]{sp::CRS}}

\item{append}{append data in new Columns}
}
\value{
tibble with transformed coordinates
}
\description{
Transform between Coordinate Reference Systems
}
\section{}{
 WSG84 (EPSG:4326) vs Web Pseudo Mercator (EPSG:3857)
See \href{https://www.maptiler.com/google-maps-coordinates-tile-bounds-projection/}{Tiles à la Google Maps} for a
detailed explanation. In Short:

EPSG:4326 uses a coordinate system on the surface of a sphere or ellipsoid of reference.
EPSG:3857 uses a coordinate system PROJECTED from the surface of the sphere or ellipsoid to a flat surface.

Usually Packages and Webservices expects coordinates for Elements in EPSG:4326, one Exception is
\code{leaflet::addRasterImage()}
}

\examples{
fn <- system.file("extdata", "meta_smn.rds", package = "rOstluft.data")
data <- readRDS(fn)
data <- dplyr::distinct(data, site, x, y)
transform(data, c(lon = "x", lat = "y"),
          sp::CRS("+init=epsg:2056"), sp::CRS("+init=epsg:4326"))

transform(data, c(lon = "x", lat = "y"),
          sp::CRS("+init=epsg:2056"), sp::CRS("+init=epsg:4326"), append = FALSE)

transfrom_LV95_to_WSG84(data)

}
\seealso{
\itemize{
\item \code{\link[sp:spTransform]{sp::spTransform()}}
\item \code{\link[sp:proj4string]{sp::proj4string()}}
\item \link[sp:CRS]{sp::CRS}
\item \href{https://proj4.org}{PROJ4}
\item \href{https://epsg.io}{epsg.io: Coordinate Systems Worldwide}
\item \href{https://rstudio.github.io/leaflet/projections.html}{Working with projections in Leaflet}
\item \href{https://www.maptiler.com/google-maps-coordinates-tile-bounds-projection/}{Tiles à la Google Maps}
}
}
