% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-projection.R
\name{transform_projection}
\alias{transform_projection}
\title{Map projection and datum transformation}
\usage{
transform_projection(df, coord = c("lon", "lat"),
  initCRS = sp::CRS("+init=epsg:4326"),
  outCRS = sp::CRS("+init=epsg:2056"))
}
\arguments{
\item{df}{data frame containing the coordinate information}

\item{coord}{columns of the coordinates. Default \code{c("lon", "lat")}}

\item{initCRS}{source coordinate reference system as \link[sp:CRS]{sp::CRS}. Default \code{sp::CRS("+init=epsg:4326")}}

\item{outCRS}{target coordinate reference system as \link[sp:CRS]{sp::CRS}. Default \code{sp::CRS("+init=epsg:2056")}}
}
\value{
data frame containing the transformed coordinatess
}
\description{
Map projection and datum transformation
}
\seealso{
\itemize{
\item \code{\link[sp:spTransform]{sp::spTransform()}}
\item \code{\link[sp:proj4string]{sp::proj4string()}}
\item \link[sp:CRS]{sp::CRS}
\item \href{https://proj4.org/}{PROJ4}
}
}
