% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic.R
\name{treshold_wrapper_function}
\alias{treshold_wrapper_function}
\title{Wrapper function for minimum capture threshold}
\usage{
treshold_wrapper_function(statistic_fun, threshold, ...)
}
\arguments{
\item{statistic_fun}{statistic function to apply if minimum capture threshold is meet}

\item{threshold}{minimum data capture threshold 0 - 1.0 to use}

\item{...}{Additional arguments passed on to the statistic function}
}
\value{
wrapped function
}
\description{
Wrapper function for minimum capture threshold
}
