% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vars}
\alias{vars}
\title{Pass grouping definitions into functions}
\usage{
vars(...)
}
\arguments{
\item{...}{Variables to group by. These arguments are automatically
\link[rlang:nse-defuse]{quoted} and later \link[rlang:eval_tidy]{evaluated} in the
context of the data frame. They support \link[rlang:nse-force]{unquoting}.}
}
\value{
named list containing quosures or symbols
}
\description{
Inspired by \code{\link[dplyr:vars]{dplyr::vars()}}, but converting strings to symbols and auto names all arguments
}
\keyword{internal}
