% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{calculate_stats}
\alias{calculate_stats}
\title{calculates a lot of stats}
\usage{
calculate_stats(data)
}
\arguments{
\item{data}{input data in rolf format as min10 or min30 data}
}
\value{
list with h1, h8gl, d1, m1, y1 data as items
}
\description{
calculates a lot of stats
}
