% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statstable.R
\name{calculate_statstable}
\alias{calculate_statstable}
\title{Calculates stats described in a table}
\usage{
calculate_statstable(data, statstable, sep = "\\\\s*,\\\\s*",
  keep_input = FALSE, order = c("input", "h1", "h8gl", "d1", "m1",
  "y1"))
}
\arguments{
\item{data}{input data in rolf format}

\item{statstable}{description of statistics to calculate in table form}

\item{sep}{seperator for combined values in statstable}

\item{keep_input}{should the input data be kept in return list as item input. Default FALSE}

\item{order}{defines the order of calculation in the from column}
}
\value{
list with one item per to interval
}
\description{
With various caveats:
\itemize{
\item data_thresh = 0.8 for all stats
\item No combining wind averaging with other stats. they must have there own row
\item to h8gl: only mean from h1
\item min10, min30, h1, d1 > y1 always use max_gap
}
}
