% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-rolf.R
\docType{class}
\name{format_rolf}
\alias{format_rolf}
\title{rOstluft long format (rolf)}
\usage{
format_rolf(tz = "Etc/GMT-1")
}
\arguments{
\item{tz}{time zone used for chunking.  Default Etc/GMT-1}
}
\value{
R6 class object of format_rolf
}
\description{
Data format for storing air quality data. A Serie is the combination of measurement site, interval,
parameter and unit.

The data is chunked in year of measurement, site and interval. Base64 encoding is used to avoid invalid characters
}
\section{Fields}{

\describe{
\item{\code{index}}{index for the data (in most cases the time columns)}

\item{\code{serie_columns}}{combination of columns for a serie}

\item{\code{chunk_columns}}{data columns for chunking}

\item{\code{chunk_calc}}{formulas for calculating new columns for chunking}

\item{\code{unique_columns}}{= index + serie_columns}

\item{\code{content_columns}}{= serie_columns + names(chunk_calc)}

\item{\code{tz}}{time zone used for chunking. Same time different chunk:
2018-01-01T00:00:00+0100 = 2018, 2017-12-31T23:00:00+0000 = 2017}
}}

\section{Methods}{


\code{$sort(data)}

\code{$merge(new_data, old_data)}

\code{$chunk_name(chunk_data)} returns the chunkname based on the first row of the supplied data

\code{$get_chunk_names(interval, site, year)} returns encoded chunknames for all combinations of the supplied vectors

\code{$encode_chunk_name(interval, site, year)} returns encoded chunkname

\code{$decode_chunk_name(chunk_name)} returns decoded chunk_name in tibble as columns (chunk_name, interval, site, year)

\code{$destroy(confirmation)} removes all files under path from the file system if "DELETE" is supplied as
confirmation
}

