% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad.R
\name{pad}
\alias{pad}
\alias{pad_serie}
\title{Pad data}
\usage{
pad(data, start_date = NULL, end_date = NULL, drop_last = FALSE)

pad_serie(serie, start_date = NULL, end_date = NULL,
  drop_last = FALSE)
}
\arguments{
\item{data}{tibble in rOstluft long format}

\item{start_date}{optional start date for padding. Default min date in series}

\item{end_date}{optional end date for padding. Default max date in series}

\item{drop_last}{optional drop the last added time point by padding. Useful when
resampling and end_date is the first time point of the new interval}

\item{serie}{a tibble in rOstluft long format containing exactly one serie}
}
\value{
tibble with padded data
}
\description{
Function to pad out missing time points
}
