% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-nabel.R
\name{read_nabel_txt}
\alias{read_nabel_txt}
\title{Read a file from Nabel}
\usage{
read_nabel_txt(fn, encoding = "latin1", tz = "Etc/GMT-1",
  interval = NULL, time_shift = NULL, na.rm = TRUE)
}
\arguments{
\item{fn}{path to input file}

\item{encoding}{encoding of the data file. Default = "latin1"}

\item{tz}{of the data. Default "Etc/GMT-1"}

\item{interval}{optional interval of the data. Use if auto detect fails. Default NULL. If used it is necessary to
define time_shift manuelly. lubridate::period(0) can be used for no shifting}

\item{time_shift}{a lubridate period to add to the time. Default NULL}

\item{na.rm}{remove na values. Default TRUE}
}
\value{
tibble in rolf format
}
\description{
Reads Exports from \href{https://www.bafu.admin.ch/bafu/en/home/topics/air/state/data/national-air-pollution-monitoring-network--nabel-.html}{National Air Pollution Monitoring Network (NABEL)}
The National Air Pollution Monitoring Network (NABEL) measures air pollution at 16 locations in Switzerland.
The stations are distributed throughout the country and monitor pollution at typical locations (e.g.
city-centre streets, residential areas, rural stations). The monitoring network has commenced operations in stages
since 1979 and is operated by the Federal office for the environment and Empa

This function reads the parameter and unit information from the header. The interval is auto detected if possible.
In Addition the time information are in end time. The time is converted to start time and the time zone defined
trough the argument tz.
The argument time_shift provides a way to manuelly shift the time series. In this case \emph{no} automatically shifting
is applied. The provided values is directly added to information in the file.
}
