% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-smn.R
\name{read_smn}
\alias{read_smn}
\title{Reads a file from the SwissMetNet of MeteoSwiss}
\usage{
read_smn(fn, tz = "Etc/GMT-1", encoding = "UTF-8",
  time_format = NULL, interval = NULL, na.rm = TRUE)
}
\arguments{
\item{fn}{path to input file}

\item{tz}{of the output data. Default "Etc/GMT-1"}

\item{encoding}{encoding of the data file. Default = "UTF-8"}

\item{time_format}{optional time_format. Use if auto detect fails. Default NULL}

\item{interval}{optional interval of the data. Use if auto detect fails. Default NULL}

\item{na.rm}{remove na values. Default TRUE}
}
\value{
tibble in rOstluft long format structure
}
\description{
\href{https://www.meteoswiss.admin.ch/home/measurement-and-forecasting-systems/land-based-stations/automatisches-messnetz.html}{SwissMetNet},
the automatic monitoring network of \href{https://www.meteoschweiz.admin.ch}{MeteoSwiss} the Federal Office for
Meteorology and Climatology, comprises about 160 automatic monitoring stations. These stations deliver a multitude
of current data on the weather and climate in Switzerland every ten minutes. The monitoring network is supplemented
by automatic precipitation stations.
}
\examples{
input <- system.file("extdata", "smn.txt", package = "rOstluft.data", mustWork = TRUE)
read_smn(single)

}
