% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample_series}
\alias{resample_series}
\title{resampling a serie}
\usage{
resample_series(
  serie,
  statistic = "mean",
  new_interval = "d1",
  data_thresh = NULL,
  max_gap = NULL,
  rename_parameter = TRUE,
  percentile = 0.95,
  skip_padding = FALSE,
  start_date = NULL,
  end_date = NULL,
  drop_last = FALSE
)
}
\arguments{
\item{serie}{a tibble in rOstluft long format containing exactly one serie}
}
\value{
tibble with resampled series
}
\description{
resampling a serie
}
\keyword{internal}
