% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample_wind}
\alias{resample_wind}
\title{Resamples wind data}
\usage{
resample_wind(data, statistic, new_interval = "d1", data.thresh = NULL,
  start.date = NULL, end.date = NULL, drop.last = FALSE,
  max_gap = NULL)
}
\value{
tibble with resampled wind data
}
\description{
Resamples wind data
}
\keyword{internal}
