% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample_wind_site}
\alias{resample_wind_site}
\title{resamples wind}
\usage{
resample_wind_site(data.site, statistic, new_interval = "d1",
  data.thresh = NULL, start.date = NULL, end.date = NULL,
  drop.last = FALSE)
}
\arguments{
\item{data.site}{wind data for one specific site. should only contain wind speed and direction}
}
\value{
tibble with resampled wind speed and direction
}
\description{
resamples wind
}
\keyword{internal}
