% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic.R
\name{get_gap_in_vector}
\alias{get_gap_in_vector}
\alias{statistic_mean}
\alias{statistic_median}
\alias{statistic_sd}
\alias{statistic_n}
\alias{statistic_sum}
\alias{statistic_max}
\alias{statistic_min}
\alias{statistic_coverage}
\alias{statistic_aot40k}
\alias{get_statistic_percentile}
\alias{get_statistic_limit_gt}
\alias{get_statistic_limit_gte}
\alias{get_statistic_limit_lt}
\alias{get_statistic_limit_lte}
\title{Get the biggest number of consecutive NA Values}
\usage{
get_gap_in_vector(x)

statistic_mean(x)

statistic_median(x)

statistic_sd(x)

statistic_n(x)

statistic_sum(x)

statistic_max(x)

statistic_min(x)

statistic_coverage(x)

statistic_aot40k(x)

get_statistic_percentile(percentile)

get_statistic_limit_gt(limit)

get_statistic_limit_gte(limit)

get_statistic_limit_lt(limit)

get_statistic_limit_lte(limit)
}
\arguments{
\item{x}{Vector}
}
\value{
number longest consecutive NA values
}
\description{
Probably not the fastest way to do it. Compared with a loop approach. The loop is a lot faster for short vectors.
But for 17500 elements (1 year min30 data) this is faster. But In the end the perfomance don't really matters.
}
\keyword{internal}
