% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stores.R
\name{store_aqmet}
\alias{store_aqmet}
\alias{store_hysplit}
\title{Predefined stores}
\usage{
store_aqmet(name = "aqmet")

store_hysplit(name = "hysplit")
}
\arguments{
\item{name}{of the store (to influence caching)}
}
\value{
store
}
\description{
\itemize{
\item \code{store_aqmet()} - Ostluft S3 storage containing airquality and meteo data
\item \code{store_hysplit()} - Ostluft S3 storage containing NOAA Hysplit trajectories
}
}
