% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmean-full-intervall.R
\name{wmean_shifted}
\alias{wmean_shifted}
\title{Calculates weighted arithmetic means from shifted time series}
\usage{
wmean_shifted(data, ..., starttime = "starttime", endtime = "endtime",
  value = "value", interval = "h1")
}
\arguments{
\item{data}{data.frame for averaging; df has to be in long format and contain a start- and end-time
column of class POSIXct (arbitrarily named)}

\item{...}{columns containing values for grouping (passed to \code{dplyr::group_by()}) when calculating
weighted means (e.g. for different measurement parameters). Columns not explicitly passed are dropped}

\item{starttime}{name of starttime column as symbol or string}

\item{endtime}{name of endtime column  as symbol or string}

\item{value}{name of column containing values to be averaged as symbol or string}

\item{interval}{specifying the output interval for averaging as string}
}
\value{
tibble with the starttime, endtime, value and grouping columns and additional the column "n" containing the
sum of weighted intervals within the averaged time interval (data availability in interval, 1 = 100%)
}
\description{
Faster special case of \code{wmean()} for evenly spaced time series. Sometimes records (e.g. measurements from miniDOAS)
provide values representative time intervals starting from odd start times to odd end times (e.g. 09:58 to 10:08 or
20 Feb to 06 March). To make such time series intercomparable and provide a standardized way of dealing with
aggregated data, \code{wmean_shifted()} provides a method to average a data.frame containing odd start- and end-time
information to standard time intervals (e.g. 10min intervals from 09:50 to 10:00) with \code{stats::weighted.mean()}.
}
\details{
sdfg
}
\section{Restrictions}{

\itemize{
\item input data has to contain evenly spaced time series (eg. 10min interval)
\item the output intervals has to be greater or equal (>=) than the input interval
\item no handling of missing data threshold. One existing value is enough
}
}

\examples{
fn <- system.file("extdata", "Zch_Stampfenbachstrasse_h1_2013_Jan.csv", package = "rOstluft.data")
data <- read_airmo_csv(fn, time_shift = lubridate::period(20, "minutes"))

df <- pluck_parameter(data, "CO") \%>\%
  pluck_unit(unit, "ppm") \%>\%
  dplyr::mutate(endtime = .data$starttime + lubridate::hours(1)) \%>\%
  dplyr::select(-interval)

wmean_shifted(df, site, parameter, unit, interval = "h1")

wmean_shifted(df, site, parameter, unit, interval = "d1")

wmean_shifted(df, site, parameter, unit, interval = "m1")
}
\seealso{
\itemize{
\item \code{rOstluft::wmean()}
}
}
\keyword{statistics}
